Three new rewriting rules. A lot of code refactoring, see below.
I started writing PreML in Preprocessed SML, which introduced some bootstrapping headaches. The result is a folder 'bootstrap' and a Bash script 'mkBootstrap.sh'. When building on a clean system a preliminary excecutable is build from the 'bootstrap' folder (which might or might not be the latest version), then a final version is build from the latest code.